feat: add ghost to RoundDetails
