+++ src/routes/main/roundDetails/NotifyDetails.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const NotifyDetails = () => {
+  return (
+    <div className="countdownDetails">
+      <div className="countdownDetails__header">
+        The next round has not opened yet.
+      </div>
+      <div className="countdownDetails__text">
+        You may still place your bids. We will notify you when the next round
+        begins
+      </div>
+    </div>
+  );
+};
+
+export default NotifyDetails;
