+++ src/routes/main/roundDetails/RoundDetails.scss
@@ -2,11 +2,6 @@
 
 .roundDetails {
   text-align: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
   .roundDetails__content {
     margin-bottom: 3rem;
     .countdownDetails {
