+++ src/routes/main/roundDetails/RoundDetailsGhost.scss
@@ -1,0 +1,3 @@
+.countdownDetails--ghost {
+  display: block !important;
+}
